+++ src/pages/favorites/favorites.tsx
@@ -1,4 +1,4 @@
-import { Helmet } from "react-helmet-async";
+import { Helmet } from 'react-helmet-async';
 
 function Favorites(): React.JSX.Element {
   return (
