+++ src/pages/login/login.tsx
@@ -1,4 +1,4 @@
-import { Helmet } from "react-helmet-async";
+import { Helmet } from 'react-helmet-async';
 
 function Login(): React.JSX.Element {
   return (
