+++ src/pages/not-found/not-found.tsx
@@ -1,4 +1,4 @@
-import { Helmet } from "react-helmet-async";
+import { Helmet } from 'react-helmet-async';
 
 function NotFound(): React.JSX.Element {
   return (
