+++ src/pages/offer/offer.tsx
@@ -1,4 +1,4 @@
-import { Helmet } from "react-helmet-async";
+import { Helmet } from 'react-helmet-async';
 
 function Offer(): React.JSX.Element {
   return (
